chore(spider): update spider.py and spider.pyi with shared RequestParamsDict
